add action yml and use inputs to make it callable
