+++ .github/actions/check-public-api/action.yml
@@ -1,0 +1,12 @@
+name: 'Check public API'
+description: 'Checks whether the public API is correctly exported.'
+inputs:
+  force-internal-exports:
+    description: 'Whether to force that everythin is exported through the internal API.'
+    default: 'false'
+  ignore-paths:
+    description: 'Paths to ignore when checking the public API.'
+
+runs:
+  using: 'node20'
+  main: 'index.js'
